Correction d'une erreur dans readLong()
git-svn-id: http://javascool.googlecode.com/svn@1218 17e3714e-118d-3e7a-01ef-bee73ecf6174

+++ work/src/org/javascool/macros/Stdin.java
@@ -177,7 +177,7 @@ public class Stdin {
       if(s.equals("")) {
         return 0;
       }
-      return readInteger(question);
+      return readLong(question);
     }
   }
   /**
